Add section for fuzzing and delta-debugging/reducing/shrinking
Fuzzing is in general the generation of data e.g. for testing functions with random values
Delta-debugging/reducing/shrinking (all three names describe the same thing but the scientific community does not agree to one) describe techniques to reduce data systematically e.g. An input X leads to an error. Give me the minimal version of X which still reproduces the error.

These two areas come normally hand in hand since the same rules for generating can be used to reduce data. I would also like to add a symbolic execution and dynamic symbolic exuction section but I do not know any tool (yet) which is written in Go and does that.